fix button margin and moved legend and covid stats to own component
